Converted from using UInt64s to UInt32s, since JavaScript doesn't _really_ support 64s.

+++ Library/HandlerRecord.cs
@@ -4,8 +4,8 @@ using System.Text;
 
 namespace InvertedTomato.WebPubSub {
     public class HandlerRecord {
-        public UInt64 TopicLow;
-        public UInt64 TopicHigh;
+        public UInt32 TopicLow;
+        public UInt32 TopicHigh;
         public Delegate Handler;
         public Type MessageType;
     }
