+++ Library/TopicRecord.cs
@@ -2,8 +2,8 @@
 
 namespace InvertedTomato.WebPubSub {
     public class TopicRecord {
-        public UInt64 Channel { get; set; }
-        public UInt64 Revision { get; set; }
+        public UInt32 Channel { get; set; }
+        public UInt32 Revision { get; set; }
         public ArraySegment<Byte> Packet { get; set; }
     }
 }
