+++ LibraryTests/EndToEndTests.cs
@@ -9,12 +9,12 @@ namespace LibraryTests {
     public class EndToEndTests {
         [Fact]
         public void EndToEnd_AfterConnection_Basic() {
-            var revisions = new Dictionary<UInt64, UInt64>();
-            var values = new Dictionary<UInt64, String>();
+            var revisions = new Dictionary<UInt32, UInt32>();
+            var values = new Dictionary<UInt32, String>();
 
             var server = new WebPubSubServer("http://+:8002/");
             var client = new WebPubSubClient("ws://localhost:8002/");
-            client.Subscribe((UInt64 topic, UInt64 revision, StringMessage message) => {
+            client.Subscribe((UInt32 topic, UInt32 revision, StringMessage message) => {
                 revisions[topic] = revision;
                 values[topic] = message.Value;
             });
@@ -30,9 +30,9 @@ namespace LibraryTests {
             Assert.Equal("Topic 0, message 1", values[0]);
             Assert.Equal("Topic 1, message 2", values[1]);
             Assert.Equal("Topic 2, message 1", values[2]);
-            Assert.Equal((UInt64)0, revisions[0]);
-            Assert.Equal((UInt64)1, revisions[1]);
-            Assert.Equal((UInt64)0, revisions[2]);
+            Assert.Equal((UInt32)0, revisions[0]);
+            Assert.Equal((UInt32)1, revisions[1]);
+            Assert.Equal((UInt32)0, revisions[2]);
 
             client.Dispose();
             Assert.True(client.IsDisposed);
@@ -42,12 +42,12 @@ namespace LibraryTests {
 
         [Fact]
         public void EndToEnd_AfterConnection_Flood() {
-            var revisions = new Dictionary<UInt64, UInt64>();
-            var values = new Dictionary<UInt64, String>();
+            var revisions = new Dictionary<UInt32, UInt32>();
+            var values = new Dictionary<UInt32, String>();
 
             var server = new WebPubSubServer("http://+:8003/");
             var client = new WebPubSubClient("ws://localhost:8003/");
-            client.Subscribe((UInt64 topic, UInt64 revision, StringMessage message) => {
+            client.Subscribe((UInt32 topic, UInt32 revision, StringMessage message) => {
                 revisions[topic] = revision;
                 values[topic] = message.Value;
             });
@@ -67,9 +67,9 @@ namespace LibraryTests {
             Assert.Equal("Topic 0, message 1", values[0]);
             Assert.Equal("Topic 1, message 50", values[1]);
             Assert.Equal("Topic 2, message 1", values[2]);
-            Assert.Equal((UInt64)0, revisions[0]);
-            Assert.Equal((UInt64)50, revisions[1]);
-            Assert.Equal((UInt64)0, revisions[2]);
+            Assert.Equal((UInt32)0, revisions[0]);
+            Assert.Equal((UInt32)50, revisions[1]);
+            Assert.Equal((UInt32)0, revisions[2]);
 
             client.Dispose();
             Assert.True(client.IsDisposed);
@@ -79,8 +79,8 @@ namespace LibraryTests {
 
         [Fact]
         public void EndToEnd_BeforeConnection_Basic() {
-            var revisions = new Dictionary<UInt64, UInt64>();
-            var values = new Dictionary<UInt64, String>();
+            var revisions = new Dictionary<UInt32, UInt32>();
+            var values = new Dictionary<UInt32, String>();
 
             var server = new WebPubSubServer("http://+:8004/");
             server.Publish(new StringMessage("Topic 0, message 1"), 0);
@@ -89,7 +89,7 @@ namespace LibraryTests {
             server.Publish(new StringMessage("Topic 2, message 1"), 2);
 
             var client = new WebPubSubClient("ws://localhost:8004/");
-            client.Subscribe((UInt64 topic, UInt64 revision, StringMessage message) => {
+            client.Subscribe((UInt32 topic, UInt32 revision, StringMessage message) => {
                 revisions[topic] = revision;
                 values[topic] = message.Value;
             });
@@ -98,9 +98,9 @@ namespace LibraryTests {
             Assert.Equal("Topic 0, message 1", values[0]);
             Assert.Equal("Topic 1, message 2", values[1]);
             Assert.Equal("Topic 2, message 1", values[2]);
-            Assert.Equal((UInt64)0, revisions[0]);
-            Assert.Equal((UInt64)1, revisions[1]);
-            Assert.Equal((UInt64)0, revisions[2]);
+            Assert.Equal((UInt32)0, revisions[0]);
+            Assert.Equal((UInt32)1, revisions[1]);
+            Assert.Equal((UInt32)0, revisions[2]);
 
             client.Dispose();
             Assert.True(client.IsDisposed);
@@ -110,8 +110,8 @@ namespace LibraryTests {
 
         [Fact]
         public void EndToEnd_BeforeConnection_Flood() {
-            var revisions = new Dictionary<UInt64, UInt64>();
-            var values = new Dictionary<UInt64, String>();
+            var revisions = new Dictionary<UInt32, UInt32>();
+            var values = new Dictionary<UInt32, String>();
 
             var server = new WebPubSubServer("http://+:8005/");
             server.Publish(new StringMessage("Topic 0, message 1"), 0);
@@ -123,7 +123,7 @@ namespace LibraryTests {
             }
 
             var client = new WebPubSubClient("ws://localhost:8005/");
-            client.Subscribe((UInt64 topic, UInt64 revision, StringMessage message) => {
+            client.Subscribe((UInt32 topic, UInt32 revision, StringMessage message) => {
                 revisions[topic] = revision;
                 values[topic] = message.Value;
             });
@@ -132,9 +132,9 @@ namespace LibraryTests {
             Assert.Equal("Topic 0, message 1", values[0]);
             Assert.Equal("Topic 1, message 50", values[1]);
             Assert.Equal("Topic 2, message 1", values[2]);
-            Assert.Equal((UInt64)0, revisions[0]);
-            Assert.Equal((UInt64)50, revisions[1]);
-            Assert.Equal((UInt64)0, revisions[2]);
+            Assert.Equal((UInt32)0, revisions[0]);
+            Assert.Equal((UInt32)50, revisions[1]);
+            Assert.Equal((UInt32)0, revisions[2]);
 
             client.Dispose();
             Assert.True(client.IsDisposed);
@@ -144,12 +144,12 @@ namespace LibraryTests {
 
         [Fact]
         public void EndToEnd_AfterConnection_Groups() {
-            var revisions = new Dictionary<UInt64, UInt64>();
-            var values = new Dictionary<UInt64, String>();
+            var revisions = new Dictionary<UInt32, UInt32>();
+            var values = new Dictionary<UInt32, String>();
 
             var server = new WebPubSubServer("http://+:8006/");
-            var client = new WebPubSubClient("ws://localhost:8006/", new UInt64[] { 1 });
-            client.Subscribe((UInt64 topic, UInt64 revision, StringMessage message) => {
+            var client = new WebPubSubClient("ws://localhost:8006/", new UInt32[] { 1 });
+            client.Subscribe((UInt32 topic, UInt32 revision, StringMessage message) => {
                 revisions[topic] = revision;
                 values[topic] = message.Value;
             });
@@ -164,7 +164,7 @@ namespace LibraryTests {
             Thread.Sleep(100);
             Assert.Single(values);
             Assert.Equal("Topic 0, message 1", values[0]);
-            Assert.Equal((UInt64)0, revisions[0]);
+            Assert.Equal((UInt32)0, revisions[0]);
 
             client.Dispose();
             Assert.True(client.IsDisposed);
